refactor(cli)!: engine-split sink commands, auto-detect mode
'sink database-changes' and 'sink from-proto' become
'sink postgres' and 'sink clickhouse'; run auto-detects the mode
from the output module type (DatabaseChanges proto -> schema.sql
mode, anything else -> relational mappings).

- engine-prefixed flags renamed: --cluster, --cursor-file-path,
  --sink-info-folder, --query-retry-count, --query-retry-sleep;
  clickhouse flags only exist on clickhouse commands
- positional <start>:<stop> range dropped; -s/--start-block and
  -t/--stop-block like 'substreams run' (inject-csv keeps its
  positional file range)
- DSN scheme validated against the command's engine
- generate-csv and inject-csv are postgres-only
