+++ docs/release-notes/change-log.md
@@ -19,7 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
-- CLI: the `substreams-sink-sql` binary is now part of the `substreams` CLI. `substreams sink database-changes {run,setup,create-user,generate-csv,inject-csv,tools}` (alias `db-changes`) covers the DatabaseChanges mode (`db_out` modules with a `schema.sql`), and `substreams sink from-proto` covers the relational-mappings mode (arbitrary protobuf output). Differences from the standalone binary: the connection string moved from a positional argument to `--dsn` (or the `SUBSTREAMS_SINK_DSN` environment variable) on every subcommand, metrics are served via the standard `--prometheus-addr` sink flag (default `localhost:9102`), pprof is an opt-in `--pprof-listen-addr` flag, and `sink database-changes tools cursor delete <module_hash>` now deletes only the given module's cursor (the standalone binary deleted all of them due to a bug).
+- CLI: the `substreams-sink-sql` binary is now part of the `substreams` CLI, organized per database engine: `substreams sink postgres {run,setup,create-user,generate-csv,inject-csv,tools}` and `substreams sink clickhouse {run,setup,create-user,tools}`. The `run` command auto-detects the mode from the selected module's output type: modules emitting `DatabaseChanges` use the `db_out` + `schema.sql` path, any other protobuf output uses the relational-mappings path. `generate-csv`/`inject-csv` are PostgreSQL-only (CSV bulk load). Differences from the standalone binary: the connection string moved from a positional argument to `--dsn` (or the `SUBSTREAMS_SINK_DSN` environment variable) on every subcommand and is validated against the command's engine (e.g. a `clickhouse://` DSN is rejected by the `postgres` commands); the block range is set with the standard `-s/--start-block` and `-t/--stop-block` sink flags (like `substreams run`) instead of a positional `<start>:<stop>` argument (`inject-csv` keeps its positional range since it selects CSV files on disk); engine-prefixed flags were renamed now that the engine is in the command name (`--clickhouse-cluster` → `--cluster`, `--clickhouse-sink-info-folder` → `--sink-info-folder`, `--clickhouse-cursor-file-path` → `--cursor-file-path`, `--clickhouse-query-retry-count` → `--query-retry-count`, `--clickhouse-query-retry-sleep` → `--query-retry-sleep`); metrics are served via the standard `--prometheus-addr` sink flag (default `localhost:9102`), pprof is an opt-in `--pprof-listen-addr` flag, and `tools cursor delete <module_hash>` now deletes only the given module's cursor (the standalone binary deleted all of them due to a bug).
 
 - Server: new `ProcessRangeRequest.merged_blocks_bundle_size` field (internal tier1→tier2 protocol) carrying the number of blocks per merged-blocks file. `0` (older tier1s) means the historical default of `100`. Tier2 applies the value per-request, so a single tier2 can serve chains with different merged-blocks sizes. **Upgrade all tier2s before setting a non-100 value on any tier1**: older tier2s ignore the field and would read the store with a bundle size of 100 (jobs stall on missing file names).
 - Server: new `Tier1Config.MergedBlocksBundleSize` (default `100`), used by tier1's own merged-blocks reads, cursor resolution, final-block rounding and forwarded to tier2 on every subrequest. The hub's kept final blocks and the live backfiller delay now scale with the bundle size.
